+++ DanielSteginkUtils/Library/ExternalFiles/GetAudioClip.cs
@@ -18,15 +18,17 @@ namespace DanielSteginkUtils.ExternalFiles
         /// <param name="fileName">Name of the file (ie. Assembly.Resources.File_Name.wav)</param>
         /// <param name="performLogging">Whether or not to log the process of getting the file</param>
         /// <returns></returns>
-        public static AudioClip? GetAudioClipFromAssembly(string assemblyName, string fileName, bool performLogging = false)
+        public static AudioClip? GetAudioClipFromAssembly(string assemblyName, string fileName, 
+                                                            bool performLogging = false)
         {
-            Logging.Log("GetAudioClipFromAssembly", $"Getting {fileName} from assembly", performLogging);
+            string logPrefix = "GetAudioClipFromAssembly";
+            Logging.Log(logPrefix, $"Getting {fileName} from assembly", performLogging);
 
             // Get the app's assembly
             Assembly assembly = Assembly.Load(assemblyName);
             if (assembly == null)
             {
-                Logging.Log("GetAudioClipFromAssembly", $"Assembly '{assemblyName}' not found", performLogging);
+                Logging.Log(logPrefix, $"Assembly '{assemblyName}' not found", performLogging);
                 return null;
             }
 
@@ -35,30 +37,44 @@ namespace DanielSteginkUtils.ExternalFiles
             {
                 if (stream == null)
                 {
-                    Logging.Log("GetAudioClipFromAssembly", $"Embedded resource '{fileName}' not found", performLogging);
+                    Logging.Log(logPrefix, $"Embedded resource '{fileName}' not found", performLogging);
                     return null;
                 }
 
-                // Parse the stream into WAV data
-                WavData wavData = new WavData();
-                if (!wavData.Parse(stream))
-                {
-                    Logging.Log("GetAudioClipFromAssembly", $"Unable to parse '{fileName}' into WAV data", performLogging);
-                    return null;
-                }
-
-                // Convert the data to an audio clip
-                float[] samples = wavData.GetSamples();
-                int sampleLength = samples.Length / wavData.FormatChunk.NumChannels;
-                int channels = wavData.FormatChunk.NumChannels;
-                int frequency = (int)wavData.FormatChunk.SampleRate;
-                AudioClip audioClip = AudioClip.Create(fileName, sampleLength, channels, frequency, false);
-                audioClip.SetData(samples, 0);
-                UnityEngine.GameObject.DontDestroyOnLoad(audioClip);
-
-                Logging.Log("GetAudioClipFromAssembly", $"'{fileName}' converted to AudioClip", performLogging);
-                return audioClip;
+                return GetAudioClipFromStream(stream, fileName, performLogging);
             }
+        }
+
+        /// <summary>
+        /// Converts a given stream into an AudioClip
+        /// </summary>
+        /// <param name="stream"></param>
+        /// <param name="clipName"></param>
+        /// <param name="performLogging"></param>
+        /// <returns></returns>
+        public static AudioClip? GetAudioClipFromStream(Stream stream, string clipName, bool performLogging = false)
+        {
+            string logPrefix = "GetAudioClipFromStream";
+
+            // Parse the stream into WAV data
+            WavData wavData = new WavData();
+            if (!wavData.Parse(stream))
+            {
+                Logging.Log(logPrefix, $"Unable to parse '{clipName}' into WAV data", performLogging);
+                return null;
+            }
+
+            // Convert the data to an audio clip
+            float[] samples = wavData.GetSamples();
+            int sampleLength = samples.Length / wavData.FormatChunk.NumChannels;
+            int channels = wavData.FormatChunk.NumChannels;
+            int frequency = (int)wavData.FormatChunk.SampleRate;
+            AudioClip audioClip = AudioClip.Create(clipName, sampleLength, channels, frequency, false);
+            audioClip.SetData(samples, 0);
+            UnityEngine.GameObject.DontDestroyOnLoad(audioClip);
+
+            Logging.Log(logPrefix, $"'{clipName}' converted to AudioClip", performLogging);
+            return audioClip;
         }
     }
 }